+++ HISTORY.md
@@ -90,6 +90,7 @@ Complete list of issues solved in the 1.0.0 milestone:
 * (#182) JobPage AJAX needs error handler to fix infinite "Updating..".
 * (#189) Shouldn't distribute runs that are being run already.
 * job.js should keep refreshing even when everything is complete and "reset" happens.
+* (#191) Preserve other window.onerror handlers (if there are any).
 
 ### Other changes